Move NetworkX policy to plugin package
