+++ DEVELOP.md
@@ -223,15 +223,3 @@ CI uses per-Python-version hashed lockfiles for supply chain security:
 - **Hash verification**: `--require-hashes` on install ensures tamper-proof installs (except AI/umap profiles where torch conflicts prevent it).
 - **Adding a dependency**: After modifying most `setup.py` extras, CI automatically regenerates artifact lockfiles. If the change affects ReadTheDocs docs dependencies, also update and commit `requirements/rtd-py3.12.lock`.
 - **Emergency override**: Set `COOLDOWN_DAYS=0` in `bin/generate-lockfiles.sh` to disable the 6-day cooldown for urgent patches.
-
-## Optional NetworkX / SciPy Policy
-
-The `networkx` extra supports `networkx>=2.5,<4`. Keep this range aligned with `graphistry.compute.networkx_policy.NETWORKX_VERSION_SPEC` and the local Cypher `graphistry.nx.*` CALL tests.
-
-SciPy is optional for NetworkX-backed calls. The `networkx-scipy` extra declares the tested SciPy range, `scipy>=1.5,<2`, for environments that want NetworkX algorithms to use SciPy-backed implementations when NetworkX chooses them. The GFQL CALL surface must still keep no-SciPy fallbacks for algorithms that have local fallbacks, such as `graphistry.nx.pagerank` and `graphistry.nx.hits`.
-
-When adding NetworkX-backed procedures:
-
-- Use `graphistry.compute.networkx_policy` for supported-version checks.
-- Update the `test-networkx-scipy-policy` CI matrix if the new procedure needs a narrower NetworkX/SciPy range.
-- Add coverage for both the lower supported NetworkX bound and the current upper-bound resolver path.
