+++ graphistry/plugins/networkx/__init__.py
@@ -1,0 +1,1 @@
+"""NetworkX optional plugin support."""
