+++ graphistry/plugins/networkx/policy.py
